refactor(group): 초대 페이지 추상화
- 현재 프로젝트에서 페이지 컴포넌트들을 만들때 세웠던 추상화 규칙에 따라 초대 페이지 컴포넌트도 추상화 진행

- styledTitle은 이미 역할이 다 드러나있고, 코드가 간단하기 때문에 따로 추상화해주지 않았음
